DEV-854 Step 2 * Cleaned up Twig HTML for step 2 related pages and partials, namely regarding correct FormValidation and FileAttach attributes on elements * Enabled show/hiding of housed by third person section * Added BIC form validation for FormValidation JS component * Added hiding AutoComplete results after delay when user blurs from input/results
I think currently that the form for step 1 isn't correctly architected regarding PHP controller and HTML layout. When I tested signing up as a "Particulier" it worked, but testing as a "Societie" did not work (PHP controller complained about not finding a "prospect_name" property) and I couldn't advance to step 2.