TECH-377 TECH-385 Fix security issue with login and lender subscription processes
As long as existing email error message was displayed, even though captcha was required,
it was possible to look for existing accounts in lender subscription process. Error
message is not displayed anymore if captcha is invalid.

Login process was broken when captcha field was not given in the request.
When captcha validation is needed, captcha code is always checked. If
field is missing, it means that captcha is not valid.